Extracted remaining androidSetup() into an AndroidSetup class
This allows us to provide release-specific Android setup routines as
there have been major changes between R13 and R14.  At the moment we are
supporting R14 which will change once we provide an R13-specific
AndroidSetup class.
