+++ src/main/groovy/com/jvoegele/gradle/plugins/android/AbstractAndroidSetup.groovy
@@ -1,0 +1,9 @@
+package com.jvoegele.gradle.plugins.android
+
+abstract class AbstractAndroidSetup implements AndroidSetup {
+  protected project
+	
+  AbstractAndroidSetup(project) {
+    this.project = project
+  }
+}
