+++ src/main/groovy/com/jvoegele/gradle/plugins/android/AndroidPlugin.groovy
@@ -39,9 +39,6 @@ class AndroidPlugin implements Plugin<Project> {
   private static final ANDROID_JARS = ['anttasks', 'sdklib', 'androidprefs', 'apkbuilder', 'jarutils']
 
   private AndroidPluginConvention androidConvention
-  private sdkDir
-  private toolsDir
-  private platformToolsDir // used since SDK r8, so check if it exists before using!
 
   private Project project
   private logger
@@ -71,56 +68,9 @@ class AndroidPlugin implements Plugin<Project> {
 	registerPropertyFiles()
 	determineAndroidDirs()
 	registerAndroidJars()
-	
-    def ant = project.ant
 
-    ant.condition('property': "exe", value: ".exe", 'else': "") { os(family: "windows") }
-    ant.condition('property': "bat", value: ".bat", 'else': "") { os(family: "windows") }
-    if (platformToolsDir.exists()) { // since SDK r8, adb is moved from tools to platform-tools
-      ant.property(name: "adb", location: new File(platformToolsDir, "adb${ant['exe']}"))
-    } else {
-      ant.property(name: "adb", location: new File(toolsDir, "adb${ant['exe']}"))
-    }
-    ant.property(name: "zipalign", location: new File(toolsDir, "zipalign${ant['exe']}"))
-    ant.property(name: 'adb.device.arg', value: '')
-
-    def outDir = project.buildDir.absolutePath
-    ant.property(name: "resource.package.file.name", value: "${project.name}.ap_")
-
-    ant.taskdef(name: 'setup', classname: 'com.android.ant.NewSetupTask', classpathref: 'android.antlibs')
-
-    // The following properties are put in place by the setup task:
-    // android.jar, android.aidl, aapt, aidl, and dx
-    ant.setup(projectTypeOut: "android.project.type",
-              androidJarFileOut: "android.jar",
-              androidAidlFileOut: "android.aidl",
-              renderScriptExeOut: "renderscript",
-              renderScriptIncludeDirOut: "android.rs",
-              bootclasspathrefOut: "android.target.classpath",
-              projectLibrariesRootOut: "project.libraries",
-              projectLibrariesJarsOut: "project.libraries.jars",
-              projectLibrariesResOut: "project.libraries.res",
-              projectLibrariesPackageOut: "project.libraries.package",
-              projectLibrariesLibsOut: "project.libraries.libs",
-              targetApiOut: "target.api")
-
-    ant.taskdef(name: "xpath", classname: "com.android.ant.XPathTask", classpathref: "android.antlibs")
-    ant.taskdef(name: "aaptexec", classname: "com.android.ant.AaptExecTask", classpathref: "android.antlibs")
-    ant.taskdef(name: "apkbuilder", classname: "com.android.ant.ApkBuilderTask", classpathref: "android.antlibs")
-
-    ant.property(name: "aapt", location: new File(platformToolsDir, "aapt${ant['exe']}"))
-    ant.property(name: "dx", location: new File(platformToolsDir, "dx${ant['bat']}"))
-
-    ant.xpath(input: androidConvention.androidManifest, expression: "/manifest/@package", output: "manifest.package")
-    // TODO: there can be several instrumentations defined
-    ant.xpath(input: androidConvention.androidManifest, expression: "/manifest/instrumentation/@android:targetPackage", output: "tested.manifest.package")
-    ant.xpath(input: androidConvention.androidManifest, expression: "/manifest/application/@android:hasCode",
-              output: "manifest.hasCode", 'default': "true")
-
-    ant.xpath(input: androidConvention.androidManifest, expression: "/manifest/instrumentation/@android:name", output: "android.instrumentation")
-    if (ant['android.instrumentation']) {
-      androidConvention.instrumentationTestsRunner = ant['android.instrumentation']
-    }
+    def setupFactory = new AndroidSetupFactory(project)
+    setupFactory.androidSetup.setup()
   }
 
   private void registerPropertyFiles() {
@@ -131,7 +81,8 @@ class AndroidPlugin implements Plugin<Project> {
 
   private void determineAndroidDirs() {
 	def ant = project.ant
-
+	def sdkDir
+	
 	// Determine the sdkDir value.
     // First, let's try the sdk.dir property in local.properties file.
     try {
@@ -153,13 +104,11 @@ class AndroidPlugin implements Plugin<Project> {
     if (sdkDir == null || sdkDir.length() == 0) {
       throw new MissingPropertyException("Unable to find location of Android SDK. Please read documentation.")
     }
-
-    toolsDir = new File(sdkDir, "tools")
-    platformToolsDir = new File(sdkDir, "platform-tools")
   }
 
   private void registerAndroidJars() {
     def ant = project.ant
+	def sdkDir = ant['sdk.dir']
 
     ant.path(id: 'android.antlibs') {
       ANDROID_JARS.each { pathelement(path: "${sdkDir}/tools/lib/${it}.jar") }
