+++ src/main/groovy/com/jvoegele/gradle/plugins/android/AndroidSetup.groovy
@@ -1,0 +1,5 @@
+package com.jvoegele.gradle.plugins.android
+
+interface AndroidSetup {
+  void setup()
+}
